chore: rename portfolio total file
